+++ hutool-core/src/main/java/cn/hutool/core/net/DefaultTrustManager.java
@@ -1,15 +1,18 @@
 package cn.hutool.core.net;
 
-import javax.net.ssl.X509TrustManager;
+import javax.net.ssl.SSLEngine;
+import javax.net.ssl.X509ExtendedTrustManager;
+import java.net.Socket;
 import java.security.cert.X509Certificate;
 
 /**
- * 默认信任管理器，默认信任所有客户端和服务端证书
+ * 默认信任管理器，默认信任所有客户端和服务端证书<br>
+ * 继承{@link X509ExtendedTrustManager}的原因见：https://blog.csdn.net/ghaohao/article/details/79454913
  *
  * @author Looly
  * @since 5.5.7
  */
-public class DefaultTrustManager implements X509TrustManager {
+public class DefaultTrustManager extends X509ExtendedTrustManager {
 
 	@Override
 	public X509Certificate[] getAcceptedIssuers() {
@@ -22,5 +25,21 @@ public class DefaultTrustManager implements X509TrustManager {
 
 	@Override
 	public void checkServerTrusted(X509Certificate[] chain, String authType) {
+	}
+
+	@Override
+	public void checkClientTrusted(X509Certificate[] x509Certificates, String s, Socket socket) {
+	}
+
+	@Override
+	public void checkServerTrusted(X509Certificate[] x509Certificates, String s, Socket socket) {
+	}
+
+	@Override
+	public void checkClientTrusted(X509Certificate[] x509Certificates, String s, SSLEngine sslEngine) {
+	}
+
+	@Override
+	public void checkServerTrusted(X509Certificate[] x509Certificates, String s, SSLEngine sslEngine) {
 	}
 }