+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * 【poi    】     CsvReader增加readFromStr（pr#1755@Github）
 * 【socket 】     SocketUtil增加connection方法
 * 【extra  】     JschUtil增加bindPort重载方法（issue#I44UTH@Github）
+* 【core   】     DefaultTrustManager改为继承X509ExtendedTrustManager
 
 ### 🐞Bug修复
 * 【core   】     改进NumberChineseFormatter算法，补充完整单元测试，解决零问题